services/agents/pyproject.toml
[project]
name = "fin-engine-agents"
version = "0.1.0"
description = "FinEngine OSS — AI agents sidecar (LLM-powered insights)"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.0",           # AWS Bedrock
    "openai>=1.50.0",          # OpenAI
    "anthropic>=0.36.0",       # Anthropic direct
    "google-genai>=1.0.0",     # Google Gemini
    "httpx>=0.27.0",           # Ollama + HTTP
    "python-dotenv>=1.0.0",    # .env loading
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
]

[project.entry-points."fin_engine.llm_providers"]
# Third-party providers register here:
# my_provider = "my_package.provider:MyProvider"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agents"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
